refactor: extract bp_overview to routes/overview.py
Phase 5.8 of the incremental dashboard.py modularisation. Moves the 7
routes registered on bp_overview — /api/channels, /api/overview,
/api/main-activity, /api/timeline, /api/cloud-cta/{status,send-otp,
verify-otp} — out of dashboard.py into routes/overview.py.

dashboard.py shrinks by 529 lines. Shared helpers (_gw_invoke,
_get_sessions, _get_crons, _get_memory_files, _find_log_file,
_infer_provider_from_model, _read_cloud_token, _write_cloud_token,
get_local_ip, SESSIONS_DIR, MEMORY_DIR, USER_NAME) stay in dashboard.py
and are accessed via late `import dashboard as _d`. The interleaved
module-level helpers (_enrich_cron_runs, _cron_runs_from_transcripts,
_find_log_file, _infer_provider_from_model) that lived between the
handlers are preserved in dashboard.py.

Pure mechanical move — zero behaviour change. Verified with smoke tests:
all 7 bp_overview endpoints return 200, /api/overview payload intact
(15 top-level keys incl. model/mainTokens/sessions/infra/system), and
8 cross-phase endpoints on other blueprints still return 200.
